Day3/Job_portal_Api/app/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
psycopg2-binary
python-jose
passlib[bcrypt]
python-dotenv
pytest
pytest-asyncio
httpx
alembic
pydantic